Enforcing np.float64 floats in test_quants

+++ tests/test_quants.py
@@ -25,12 +25,12 @@ def test_weighted_mean():
 
 
 def test_weighted_std():
-    x = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9])
-    y = np.array([9, 8, 7, 6, 5, 4, 3, 2, 1])
+    x = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9]).astype(np.float64)
+    y = np.array([9, 8, 7, 6, 5, 4, 3, 2, 1]).astype(np.float64)
     Sigma = np.cov(x, y)
-    weights = np.array([1, 1])
+    weights = np.array([1, 1]).astype(np.float64)
     assert weighted_std(Sigma, weights) == 0.0
-    weights = np.array([-3, 5])
+    weights = np.array([-3, 5]).astype(np.float64)
     assert weighted_std(Sigma, weights) ** 2 == 480.0
 
 
@@ -77,26 +77,26 @@ def test_value_at_risk_invalid_types():
 
 
 def test_annualised_portfolio_quantities():
-    x = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9])
-    y = np.array([9, 8, 7, 6, 5, 4, 3, 2, 1])
+    x = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9]).astype(np.float64)
+    y = np.array([9, 8, 7, 6, 5, 4, 3, 2, 1]).astype(np.float64)
     Sigma = np.cov(x, y)
-    mean = np.array([1, 2])
-    weights = np.array([-3, 5])
-    res = annualised_portfolio_quantities(weights, mean, Sigma, 0, 252)
+    mean = np.array([1, 2]).astype(np.float64)
+    weights = np.array([-3, 5]).astype(np.float64)
+    res = annualised_portfolio_quantities(weights, mean, Sigma, 0.0, 252)
     orig = (1764.0, 347.79304190854657, 5.071981861166303)
     for i in range(len(res)):
         assert abs(res[i] - orig[i]) <= 1e-15
 
 
 def test_downside_risk():
-    data1 = pd.DataFrame({"1": [1, 2, 4, 8], "2": [1, 2, 3, 4]})
-    weights = np.array([0.25, 0.75])
+    data1 = pd.DataFrame({"1": [1, 2, 4, 8], "2": [1, 2, 3, 4]}).astype(np.float64)
+    weights = np.array([0.25, 0.75]).astype(np.float64)
     rf_rate = 0.005
     dr1 = downside_risk(data1, weights, rf_rate)
     assert dr1 == 0
 
-    data2 = pd.DataFrame({"1": [7, 6, 5, 4, 3]})
-    weights = np.array([1])
+    data2 = pd.DataFrame({"1": [7, 6, 5, 4, 3]}).astype(np.float64)
+    weights = np.array([1]).astype(np.float64)
     rf_rate = 0.0
     dr2 = downside_risk(data2, weights, rf_rate)
     assert abs(dr2 - 0.19409143531019335) <= 1e-15
